[ll] Adding GPIO autosetting for ADC.

+++ scm_v3c/scm3C_hardware_interface.c
@@ -184,7 +184,7 @@ void GPI_enables(unsigned int mask){
 	}
 }
 
-unsigned short get_GPI_enables() {
+unsigned int get_GPI_enables() {
 	/*
 	Inputs:
 		None.
